user-service/requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
djangorestframework-simplejwt>=5.3.0,<5.4
drf_spectacular>=0.15.1,<0.16